+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added support for custom unauthorized response
 - Added dynamic json responses
+- Dynamic Route payload is validated upon creation
 
 ## [2.0.2] - 2021-04-23
 ### Fixed